requirements.txt
discord.py
aiohttp
croniter
pytz
psutil
gputil
discord-ext-menus
platform